Rename reef scheduler to yarn scheduler
REEF is a framework to develop YARN applications. Using cluster type, YARN in
this case, for name of scheduler instead of name of the framework used.
