+++ heron/config/src/yaml/conf/yarn/scheduler.yaml
@@ -1,8 +1,8 @@
 # scheduler class for distributing the topology for execution
-heron.class.scheduler:                       com.twitter.heron.scheduler.reef.ReefScheduler
+heron.class.scheduler:                       com.twitter.heron.scheduler.yarn.YarnScheduler
 
 # launcher class for submitting and launching the topology
-heron.class.launcher:                        com.twitter.heron.scheduler.reef.ReefLauncher
+heron.class.launcher:                        com.twitter.heron.scheduler.yarn.YarnLauncher
 
 # working directory for the topologies
 heron.scheduler.local.working.directory:     ${HOME}/.herondata/topologies/${CLUSTER}/${ROLE}/${TOPOLOGY}
