+++ heron/schedulers/src/java/com/twitter/heron/scheduler/yarn/HeronClientConfiguration.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.twitter.heron.scheduler.reef;
+package com.twitter.heron.scheduler.yarn;
 
 import org.apache.reef.client.ClientConfiguration;
 import org.apache.reef.tang.formats.ConfigurationModule;
