+++ tools/rules/heron_client.bzl
@@ -21,9 +21,9 @@ def heron_client_aurora_files():
         "//heron/config/src/yaml:conf-aurora-yaml",
     ]
 
-def heron_client_reef_files():
+def heron_client_yarn_files():
     return [
-        "//heron/config/src/yaml:conf-reef-yaml",
+        "//heron/config/src/yaml:conf-yarn-yaml",
     ]
 
 def heron_client_lib_scheduler_files():
@@ -32,7 +32,7 @@ def heron_client_lib_scheduler_files():
         "//heron/schedulers/src/java:heron-local-scheduler",
         "//heron/schedulers/src/java:heron-aurora-scheduler",
         "//heron/schedulers/src/java:heron-slurm-scheduler",
-        "//heron/schedulers/src/java:heron-reef-scheduler",
+        "//heron/schedulers/src/java:heron-yarn-scheduler",
         "//heron/packing/src/java:heron-roundrobin-packing",
     ]
